brocolib/utils/pyproject.toml
[tool.poetry]
name = "brocolib-utils"
version = "0.0.35"
description = "Utilities to use Brocoli Data Platform"
authors = ["Brocoli Data, <contact.brocoli@gmail.com>"]
readme = "README.md"
packages = [{include = "brocolib_utils"}]

[tool.poetry.dependencies]
python = "^3.8"
gspread = "^5.7.2"
pandas = "^1.5.2"
pandas-gbq = "^0.18.1"
oauth2client = "^4.1.3"
ruamel-yaml = "^0.17.21"
google-api-python-client = "^2.70.0"
google-cloud-storage = "^2.7.0"
pyyaml = "^6.0"
gspread-pandas = "^3.2.2"
fsspec = "^2022.11.0"
gcsfs = "^2022.11.0"
typer = {extras = ["all"], version = "^0.7.0"}


[tool.poetry.group.dev.dependencies]
pytest = "^7.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
ddm = "brocolib_utils.ddm.cli:app"
fast_dbt = "brocolib_utils.fast_dbt.cli:app"
